+++ UXToolsGame/Plugins/UXTools/Source/UXTools/Private/Interactions/UxtManipulatorComponentBase.cpp
@@ -3,7 +3,6 @@
 
 #include "Interactions/UxtManipulatorComponentBase.h"
 
-#include "Constraints/UxtConstraintManager.h"
 #include "Engine/World.h"
 #include "Interactions/Manipulation/UxtManipulationMoveLogic.h"
 #include "Interactions/Manipulation/UxtTwoHandRotateLogic.h"
@@ -140,8 +139,6 @@ void UUxtManipulatorComponentBase::SetInitialTransform()
 
 	FTransform headPose = UUxtFunctionLibrary::GetHeadPose(GetWorld());
 	InitialCameraSpaceTransform = InitialTransform * headPose.Inverse();
-
-	Constraints->Initialize(InitialTransform);
 }
 
 void UUxtManipulatorComponentBase::ApplyTargetTransform(const FTransform& TargetTransform)
@@ -169,20 +166,11 @@ void UUxtManipulatorComponentBase::BeginPlay()
 		OnBeginGrab.AddDynamic(this, &UUxtManipulatorComponentBase::OnManipulationStarted);
 		OnEndGrab.AddDynamic(this, &UUxtManipulatorComponentBase::OnManipulationEnd);
 	}
-
-	Constraints = new UxtConstraintManager(*GetOwner());
 }
 
 void UUxtManipulatorComponentBase::EndPlay(const EEndPlayReason::Type EndPlayReason)
 {
-	delete Constraints;
 	Super::EndPlay(EndPlayReason);
-}
-
-void UUxtManipulatorComponentBase::TickComponent(float DeltaTime, ELevelTick TickType, FActorComponentTickFunction* ThisTickFunction)
-{
-	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
-	Constraints->Update(TransformTarget->GetComponentTransform());
 }
 
 void UUxtManipulatorComponentBase::OnManipulationStarted(UUxtGrabTargetComponent* Grabbable, FUxtGrabPointerData GrabPointer)