+++ UXToolsGame/Plugins/UXTools/Source/UXTools/Public/Interactions/UxtManipulationFlags.h
@@ -11,21 +11,11 @@
 UENUM(BlueprintType)
 enum class EUxtOneHandRotationMode : uint8
 {
-	/** Does not rotate object as it is being moved. */
-	MaintainOriginalRotation,
 	/** Only works for articulated hands/controllers. Rotate object using rotation of the hand/controller, but about the object center
 	   point. Useful for inspecting at a distance. */
 	RotateAboutObjectCenter,
 	/** Only works for articulated hands/controllers. Rotate object as if it was being held by hand/controller. Useful for inspection. */
 	RotateAboutGrabPoint,
-	/** Maintains the object's original rotation for Y/Z axis to the user. */
-	MaintainRotationToUser,
-	/** Maintains object's original rotation to user, but makes the object vertical. Useful for bounding boxes. */
-	GravityAlignedMaintainRotationToUser,
-	/** Ensures object always faces the user. Useful for slates/panels. */
-	FaceUser,
-	/** Ensures object always faces away from user. Useful for slates/panels that are configured backwards. */
-	FaceAwayFromUser,
 };
 
 /** Two-handed transformations supported by the generic manipulator. */